requirement.txt
google-generativeai
load_dotenv